Sarancha_Lesson_6_work | Создал новую ветку
