+++ app/src/main/java/ru/aston/sarancha_lesson2/MainActivity.kt
@@ -3,10 +3,10 @@ package ru.aston.sarancha_lesson2
 import android.os.Bundle
 import android.view.View
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.isNotEmpty
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import kotlinx.android.synthetic.main.fragment_main.*
+import ru.aston.sarancha_lesson2.view.*
 import ru.aston.sarancha_lesson2.Utils.makeGone
 import ru.aston.sarancha_lesson2.Utils.makeVisible
 import ru.aston.sarancha_lesson2.contract.HasCustomTitle
