+++ app/src/main/java/ru/aston/sarancha_lesson2/view/OfficeInfoFragment.kt
@@ -1,10 +1,12 @@
-package ru.aston.sarancha_lesson2
+package ru.aston.sarancha_lesson2.view
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import ru.aston.sarancha_lesson2.OfficeAddress
+import ru.aston.sarancha_lesson2.R
 import ru.aston.sarancha_lesson2.contract.HasCustomTitle
 import ru.aston.sarancha_lesson2.databinding.FragmentOfficeInfoBinding
